refactor(ReusableComponents): rename to singular, remove multiselect
