+++ inc/fieldGroups/pageComponents.php
@@ -16,17 +16,17 @@ add_action('Flynt/afterRegisterComponents', function () {
                 'button_label' => __('Add Component', 'flynt'),
                 'layouts' => [
                     Components\BlockCollapse\getACFLayout(),
-                    Components\BlockImage\getACFLayout(),
                     Components\BlockImageText\getACFLayout(),
+                    Components\BlockImage\getACFLayout(),
                     Components\BlockVideoOembed\getACFLayout(),
                     Components\BlockWysiwyg\getACFLayout(),
                     Components\GridImageText\getACFLayout(),
                     Components\GridPostsLatest\getACFLayout(),
                     Components\ListComponents\getACFLayout(),
-                    Components\SliderImages\getACFLayout(),
                     Components\ReusableComponents\getACFLayout(),
+                    Components\SliderImages\getACFLayout(),
                 ],
-            ]
+            ],
         ],
         'location' => [
             [
